+++ app/Http/Controllers/UsuarioController.php
@@ -6,29 +6,41 @@ use App\Models\Usuario;
 use App\Models\Rol;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use App\Http\Controllers\BitacoraController;
+use App\Models\Bitacora;
 
 class UsuarioController extends Controller
 {
 
     //vista principal de gestión de usuarios
     public function index()
-    {
-        $usuarios = Usuario::with('rol')->get(); // relacionar con el rol
+    {   
+        $this->authorize('viewAny', Usuario::class); // Solo si tiene permiso 'Ver Usuarios'
+        $usuarios = Usuario::with('rol')->paginate(10); // relacionar con el rol
         return view('pages.gestion.usuarios.index', [
             'usuarios' => $usuarios,
             'eliminados' => false
         ]);
     }
-
+      //vista principal de gestión de usuarios
+    public function index2()
+    {   
+        $this->authorize('viewAny', Usuario::class); // Solo si tiene permiso 'Ver Usuarios'
+        $usuarios = Usuario::paginate(10); // relacionar con el rol
+        return view('pages.gestion.bitacora.index', [
+            'usuarios' => $usuarios,
+            'eliminados' => false
+        ]);
+    }
     //vista formulario de registro de usuario
     public function create()
-    {
+    {   $this->authorize('create', Usuario::class); // Solo si tiene permiso 'Agregar Usuarios'
         $roles = Rol::all(); // Obtener todos los roles
         return view('pages.gestion.usuarios.create', compact('roles'));
     }
     //guarda usuario nuevo
     public function store(Request $request)
-    {
+    {   $this->authorize('create', Usuario::class); // Protección también desde el backend
         // Validación de los campos del formulario
         $request->validate([
             'nombre_usuario' => 'required|string|max:100',
@@ -44,15 +56,18 @@ class UsuarioController extends Controller
         $usuario->password_usuario = $request->password_usuario; //Se encripta automáticamente en el modelo
         $usuario->id_rol = $request->id_rol;
         $usuario->save();
-
+        BitacoraController::registrar(
+            'CREAR',
+            'Se registró un nuevo usuario: ' . $usuario->nombre_usuario
+        );
         return redirect()->route('usuario.index')->with('success', 'Usuario administrador registrado correctamente.');
     }
 
 
 
     public function eliminados()
-    {
-        $usuarios = Usuario::onlyTrashed()->with('rol')->get();
+    {   $this->authorize('viewAny', Usuario::class); // Protege ver eliminados también
+        $usuarios = Usuario::onlyTrashed()->with('rol')->paginate(10);
         return view('pages.gestion.usuarios.index', [
             'usuarios' => $usuarios,
             'eliminados' => true
@@ -65,6 +80,7 @@ class UsuarioController extends Controller
     public function edit($id_usuario)
     {
         $usuario = Usuario::findOrFail($id_usuario);
+         $this->authorize('update', $usuario); // Solo si tiene permiso 'Editar Usuarios'
         $roles = Rol::all(); // Obtener roles para el formulario
         return view('pages.gestion.usuarios.edit', compact('usuario', 'roles'));
     }
@@ -75,6 +91,9 @@ class UsuarioController extends Controller
      */
     public function update(Request $request, $id_usuario)
     {
+        // Buscar el usuario por ID
+        $usuario = Usuario::findOrFail($id_usuario);
+        $this->authorize('update', $usuario); // Protección en update
         // Validación
         $request->validate([
             'nombre_usuario' => 'required|string|max:255',
@@ -82,15 +101,17 @@ class UsuarioController extends Controller
             'id_rol' => 'required|exists:rols,id_rol',
         ]);
 
-        // Buscar el usuario por ID
-        $usuario = Usuario::findOrFail($id_usuario);
+
 
         // Actualizar datos
         $usuario->nombre_usuario = $request->nombre_usuario;
         $usuario->correo_usuario = $request->correo_usuario;
         $usuario->id_rol = $request->id_rol;
         $usuario->save();
-
+        BitacoraController::registrar(
+            'ACTUALIZAR',
+            'Se actualizó el usuario: ' . $usuario->nombre_usuario
+        );
         return redirect()->route('usuario.index')->with('success', 'Usuario actualizado correctamente.');
     }
 
@@ -101,8 +122,12 @@ class UsuarioController extends Controller
     public function destroy($id_usuario)
     {
         $usuario = Usuario::findOrFail($id_usuario);
+        $this->authorize('delete', $usuario); // Solo si tiene permiso 'Eliminar Usuarios'
         $usuario->delete(); // Eliminación lógica (SoftDeletes)
-
+        BitacoraController::registrar(
+            'ELIMINAR',
+            'Se eliminó el usuario: ' . $usuario->nombre_usuario
+        );
         return redirect()->route('usuario.index')->with('success', 'Usuario eliminado correctamente.');
     }
 
@@ -112,6 +137,7 @@ class UsuarioController extends Controller
     public function restore($id_usuario)
     {
         $usuario = Usuario::withTrashed()->findOrFail($id_usuario);
+        $this->authorize('restore', $usuario); // Solo si tiene permiso para restaurar
         $usuario->restore(); // Restaurar el usuario
 
         return redirect()->route('usuario.index')->with('success', 'Usuario restaurado correctamente.');
